infra: move locale-proxy to internal
